Add controllers for types
